Fix gyst suite bug
Bug: Signup account not appearing in the gyst suite "new service sestting
editor".

Reason: Was using oauth_connected_user_entry_id instead of the user_id
when logging in

Fix: Propeprly use user_id.

---

Bug2: A new oauth_connected_user, `is_sign_up:false` gets created with
the same service_id I used for signing up but with different user_id.

Reason: Side effect of the bug 1. Because sign up was done with
oauth_connected_user_entry_id instead of user_id, when connecting the
service to add the github account into the /suite page, the server
understands it as connecting a new service. Again, the user has wrongly
logged in with oauth_connectted_user_entry_id, so the user is logged in
from the perspective of the server.

Fix: Same fix as bug1.