Fix Vuex typescript bug
Vuex typescript wasn't working when it did in brand new nuxt apps.

Took SOOO LONG to fix this only to find out that the fix was changing
if-else statement for `nuxt.ready` and `builder.build()` which WAS the
default code for running nuxt.

The brand new nuxt app now has `await nuxt.read()` run before if
statement and there is no else statement.

Tried updating modules, so the `package.json` and `package-lock.json` are
updated.

`service-setting/index.ts` was updated because it started throwing error
regarding Int32 not being a number.