Update pagination method
Pagination is done more simply. It ignores datetime but depends on how
many entries exist in the proloaded entries instead.

And pagination request is made automatically when the number of entries
for that 'param instance' is under certain amount of number.

Decided to choose this way because it's a lot simpler and just works.