WIP
Lots of fixes and deleted many components.

Last worked on (2020-06-08 night):

Need to work on fixing automatically making pagination request when there
is a 'rate limit' error. For example, I want this warning to trigger
making pagination request when clicking on 'load more' or
'scroll to bottom', and not when the client receives response. Infact,
the current implementation repeatedly makes request when the response
contains rate limit error.