Handle rate limit more properly
I think this is the major progress in this commit.

Next todo is to handle `end of response` (eg, after getting the last
commit for github repo).

Currently, this has the same behavior as how rate_limit was poorly
handled; repeatedly making the request and sending response with same
warning.