Drop pagination_data.index
Dropped this propeprty because it's everywhere but not used. And this
can be mimicked if a pagination value like this is needed just like in
github. `{ new: 2, old: null }` then in the next 'old' pagination request
the "pagination data" (previously 'pagination_data.options') will be
`{ new: 3, old: 1 }`.

I thought this was necessary for implementing "ALL_ENTRIES_LOADED", but I
forgot why.

But this change works.