Add warning to Trello & remove unused function
Unused function in loader, `getPaginationReqData`. It's replaced with
`getPaginationReqDataForLoadEntryParam` and iterating the `load status`
when having to consider all pagination for all load status, as done in
the function `loadMore` in `pages/index.vue`.

Adding warnign to Trello fixed some errors as well. It relies on id of
the returned entries, but when it returns an empty array, it results in
"Key Error. Cannot access 'id' of undefined" or something.

So, this should be applied to other services that rely on id of returned
entries.