Remove update connected account feature
It was a "someone might need this" feature. This requires more works such
as having to consider a case where the user is updating the connected
user for service setting that uses setting values. The setting values may
require authorized access but after updating the user, the setting values
may not work because the new user is not authorized to that resource on
the outside service side. For example, private repo of a github user.