Add revoke dialog
Some changes were made so that the dialog can inform the user about
what kind of 'setting' will be removed.

A new function `refreshTokenIfFailOAuthServiceId` had to be created a new
because this was the simplest solution I could come up without having to
create `getOAuthServiceId` from a service id used in the suite.
Apparently, the `oauth_service_id` is used for `/settings/accounts` so it
can be directly used as the 'service_id'. This was causing problem for
facebook, bitbucket both of which the loader module for `getServiceInfo`
wasn't defined, and for `google` in case which just 'google-calendar' and
the 'youtube' are registered but no 'google'.

However the change doesn't remove the service setting just yet.