Fix sign up process
After the user signs up and gets redirected to `/login` page, it didn't
render the service buttons but just left it empty, and with the
`relogin=true` parameter, it was annoying to see the dialog twice, once
after redirected, second after refreshing the page.

And this fix is how it's supposed to work anyways.