Update pagination loading behavior & fix bug
The previous pagination load behavior made loop of requests when the
response had 0 entries, and the pagination was done in date range. The
specific case is the google calendar. This is annoying and unnecessary.

In such case, the pagination request will be made only when "load more"
action is triggered by scrolling to the bottom of the page or clicking on
"load more".

And I thought a bug was related to google calendar, but was a github bug.
Fix this bug and added a utility function.