Make handler instance for each request
This is the way it's SUPPOSED to be. Part of 2020-w28-todo5. Will be
finished when I rollback the websocket handler part.

One way to see what the problem is, use the get-google-calendars request
handler. Print `this.res_data` before assigning at line 28.

On the first request, it's empty, on the next request, it has value. And
that is to say that state is shared among all users.